Implemented operations for question, debug email functionaloty

+++ server/controller/queCategory.js
@@ -5,7 +5,8 @@ const {wrapAsync} = require("../utils/wrapAsync");
 
 const fetchCategory = wrapAsync(async (req, res) => {
     const categories = await queCategoryModel.find({});
-    res.json(categories);
+    const categoryName = categories.map((category) => category.categoryName);
+    res.json(categoryName);
 });
 
 const addCategory = wrapAsync(async (req, res) => {
@@ -17,6 +18,13 @@ const addCategory = wrapAsync(async (req, res) => {
 
 const removeCategory = wrapAsync(async (req, res) => {
     const {category} = req.params;
+    //delete all question for that category
+    const categoryObj = await queCategoryModel.findOne({categoryName: category});
+    if (!categoryObj) {
+        return res.status(404).json({msg: "Category not found"});
+    }
+    const questions = await queModel.deleteMany({category: categoryObj._id});
+    //delete the category
     await queCategoryModel.deleteOne({categoryName: category});
     res.json({msg: "Category removed successfully."});
 });
